+++ area-restriction-service/src/main/java/com/comit/services/areaRestriction/business/AreaRestrictionManagerNotificationBusinessImpl.java
@@ -2,15 +2,21 @@ package com.comit.services.areaRestriction.business;
 
 import com.comit.services.areaRestriction.client.data.EmployeeDtoClient;
 import com.comit.services.areaRestriction.client.data.LocationDtoClient;
+import com.comit.services.areaRestriction.constant.AreaRestrictionErrorCode;
+import com.comit.services.areaRestriction.constant.Const;
 import com.comit.services.areaRestriction.controller.request.ManagerTimeSkip;
+import com.comit.services.areaRestriction.exception.AreaRestrictionCommonException;
 import com.comit.services.areaRestriction.model.entity.AreaRestrictionManagerNotification;
 import com.comit.services.areaRestriction.service.AreaRestrictionManagerNotificationService;
 import com.comit.services.areaRestriction.service.AreaRestrictionServices;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class AreaRestrictionManagerNotificationBusinessImpl implements AreaRestrictionManagerNotificationBusiness {
@@ -19,6 +25,10 @@ public class AreaRestrictionManagerNotificationBusinessImpl implements AreaRestr
     private AreaRestrictionManagerNotificationService areaRestrictionManagerNotificationService;
     @Autowired
     private AreaRestrictionServices areaRestrictionServices;
+    @Autowired
+    private HttpServletRequest httpServletRequest;
+    @Value("${app.internalToken}")
+    private String internalToken;
 
     @Override
     public List<AreaRestrictionManagerNotification> saveAreaManagerTimeList(List<ManagerTimeSkip> managerTimeSkips, Integer areaRestrictionId) {
@@ -50,6 +60,11 @@ public class AreaRestrictionManagerNotificationBusinessImpl implements AreaRestr
 
     @Override
     public boolean deleteARManagerNotificationOfEmployee(Integer employeeId) {
+        if (!isInternalFeature()) throw new AreaRestrictionCommonException(AreaRestrictionErrorCode.PERMISSION_DENIED);
         return areaRestrictionManagerNotificationService.deleteAreaRestrictionManagerNotificationListOfEmployee(employeeId);
+    }
+
+    public boolean isInternalFeature() {
+        return Objects.equals(httpServletRequest.getHeader("token"), internalToken);
     }
 }
